fileUrl

+++ frameworks/native/player/player_impl.cpp
@@ -1718,7 +1718,7 @@ int32_t PlayerImpl::OpenFile(const std::string& fileName)
     CHECK_AND_RETURN_RET_NOLOG(ret == MSERR_OK, ret);
     int fd = open(fileName.c_str(), O_RDONLY);
     CHECK_AND_RETURN_RET_NOLOG(fd != -1, MSERR_INVALID_VAL);
-    int64_t fileSize = static_cast<int64_t>(GetFileSize(fileName));
+    int64_t fileSize = GetFileSize(fileName);
     if (!fdsanFd_) {
         fdsanFd_ = std::make_unique(FdsanFd)(fd);
     } else {
@@ -1743,13 +1743,13 @@ int32_t PlayerImpl::CheckFileStat(const std::string& fileName)
     return MSERR_OK;
 }
 
-size_t PlayerImpl::GetFileSize(const std::string& fileName)
+int64_t PlayerImpl::GetFileSize(const std::string& fileName)
 {
-    size_t fileSize = 0;
+    int64_t fileSize = 0;
     if (!fileName.empty()) {
         struct stat fileStatus {};
         if (stat(fileName.c_str(), &fileStatus) == 0) {
-            fileSize = static_cast<size_t>(fileStatus.st_size);
+            fileSize = static_cast<int64_t>(fileStatus.st_size);
         }
     }
     return fileSize;
